Further cleanup and remove more benchmark code, consolidating as needed
